+++ src/project/cookbook.rs
@@ -1,15 +1,25 @@
 use std::{collections::BTreeMap, path::PathBuf};
 
-use crate::project::Recipe;
+use crate::{
+    project::Recipe,
+    template::{parse_template, parse_variable_list},
+};
 use anyhow::bail;
 use ignore::WalkBuilder;
+use indexmap::IndexMap;
 use log::debug;
 use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 pub struct Cookbook {
     pub name: String,
-    pub variables: Option<Vec<String>>,
+
+    #[serde(default)]
+    pub environment: Vec<String>,
+
+    #[serde(default)]
+    pub variables: IndexMap<String, String>,
+
     pub recipes: BTreeMap<String, Recipe>,
 
     #[serde(skip)]
@@ -19,9 +29,13 @@ impl Cookbook {
     /// Creates a cookbook config from a path to a cookbook file or directory
     ///
     /// # Arguments
-    /// * `path` - Path to a cookbook file or directory containing a cookbook.ya?ml file
+    /// * `path` - Path to a cookbook file or directory containing a `cookbook.ya?ml` file
     ///
-    pub fn from(path: &PathBuf) -> anyhow::Result<Self> {
+    pub fn from(
+        path: &PathBuf,
+        project_environment: &[String],
+        project_variables: &IndexMap<String, String>,
+    ) -> anyhow::Result<Self> {
         let config: Cookbook;
 
         let config_str = match std::fs::read_to_string(path) {
@@ -32,10 +46,40 @@ impl Cookbook {
         match serde_yaml::from_str::<Self>(&config_str) {
             Ok(mut parsed) => {
                 parsed.config_path = path.to_path_buf();
-                parsed.recipes.iter_mut().for_each(|(name, recipe)| {
+
+                // Inherit environment and variables from project
+                let mut cookbook_environment = project_environment.to_owned();
+                cookbook_environment.extend(parsed.environment.iter().cloned());
+                parsed.environment = cookbook_environment;
+
+                let mut cookbook_variables = project_variables.clone();
+                cookbook_variables.extend(parsed.variables.clone());
+
+                parsed.variables = parse_variable_list(&parsed.environment, &cookbook_variables)?;
+
+                parsed.recipes.iter_mut().try_for_each(|(name, recipe)| {
                     recipe.name = name.clone();
                     recipe.cookbook = parsed.name.clone();
                     recipe.config_path = path.to_path_buf();
+
+                    // Inherit environment and variables from cookbook
+                    let mut recipe_environment = parsed.environment.clone();
+                    recipe_environment.extend(recipe.environment.iter().cloned());
+                    recipe.environment = recipe_environment;
+
+                    let mut recipe_variables = parsed.variables.clone();
+                    recipe_variables.extend(recipe.variables.clone());
+                    if let Ok(variables) =
+                        parse_variable_list(recipe.environment.as_slice(), &recipe_variables)
+                    {
+                        recipe.variables = variables;
+                    } else {
+                        bail!("Could not parse recipe variables: {}", recipe.name)
+                    }
+
+                    recipe.run =
+                        parse_template(&recipe.run, &recipe.environment, &recipe.variables)?;
+
                     if let Some(dependencies) = recipe.dependencies.as_ref() {
                         let new_deps = dependencies.iter().map(|dep| {
                             if !dep.contains(':') {
@@ -46,7 +90,9 @@ impl Cookbook {
                         });
                         recipe.dependencies = Some(new_deps.collect());
                     }
-                });
+
+                    Ok(())
+                })?;
                 config = parsed;
             }
             Err(err) => bail!("Could not parse cookbook file: {}", err),
@@ -57,13 +103,17 @@ impl Cookbook {
 
     /// Gets all cookbooks recursively in a directory
     ///
-    /// map_from recursively searches for all cookbooks in a directory respecting .gitignore and
-    /// .bakeignore files
+    /// map_from recursively searches for all cookbooks in a directory respecting `.gitignore` and
+    /// `.bakeignore` files
     ///
     /// # Arguments
     /// * `path` - Path to a directory
     ///
-    pub fn map_from(path: &PathBuf) -> anyhow::Result<BTreeMap<String, Self>> {
+    pub fn map_from(
+        path: &PathBuf,
+        project_environment: &[String],
+        project_variables: &IndexMap<String, String>,
+    ) -> anyhow::Result<BTreeMap<String, Self>> {
         let all_files = WalkBuilder::new(path)
             .add_custom_ignore_filename(".bakeignore")
             .build();
@@ -72,7 +122,8 @@ impl Cookbook {
                 Ok(file) => {
                     let filename = file.file_name().to_str().unwrap();
                     if filename.contains("cookbook.yaml") || filename.contains("cookbook.yml") {
-                        match Self::from(&file.into_path()) {
+                        match Self::from(&file.into_path(), project_environment, project_variables)
+                        {
                             Ok(cookbook) => Some(Ok((cookbook.name.clone(), cookbook))),
                             Err(err) => Some(Err(err)),
                         }
@@ -92,6 +143,7 @@ impl Cookbook {
 mod test {
     use std::{collections::BTreeMap, path::PathBuf};
 
+    use indexmap::IndexMap;
     use test_case::test_case;
 
     fn config_path(path_str: &str) -> String {
@@ -99,7 +151,7 @@ mod test {
     }
 
     fn validate_cookbook_foo(actual: anyhow::Result<super::Cookbook>) {
-        assert_eq!(actual.unwrap().name, "foo")
+        assert_eq!(actual.as_ref().unwrap().name, "foo");
     }
 
     fn validate_cookbook_vec(actual: anyhow::Result<BTreeMap<String, super::Cookbook>>) {
@@ -110,12 +162,12 @@ mod test {
     #[test_case(config_path("/invalid/config/cookbook.yml") => matches Err(_); "Invalid cookbook file")]
     #[test_case(config_path("/invalid/config") => matches Err(_); "Cant read directory")]
     fn read_cookbook(path_str: String) -> anyhow::Result<super::Cookbook> {
-        super::Cookbook::from(&PathBuf::from(path_str))
+        super::Cookbook::from(&PathBuf::from(path_str), &[], &IndexMap::new())
     }
 
     #[test_case(config_path("/valid/") => using validate_cookbook_vec; "Root dir")]
     #[test_case(config_path("/invalid/config") => matches Err(_); "Invalid dir")]
     fn read_all_cookbooks(path_str: String) -> anyhow::Result<BTreeMap<String, super::Cookbook>> {
-        super::Cookbook::map_from(&PathBuf::from(path_str))
+        super::Cookbook::map_from(&PathBuf::from(path_str), &[], &IndexMap::new())
     }
 }